# docs/contracts/openapi/transition_service-openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Transition Service",
    "version": "1.0.0"
  },
  "paths": {
...
    "/early-warning": {
      "post": {
        "summary": "Early Warning",
        "operationId": "early_warning_early_warning_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EarlyWarningRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StandardResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/exposure": {
      "post": {
        "summary": "Occupation Exposure",
        "operationId": "occupation_exposure_exposure_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ExposureRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StandardResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/skills-gap": {
      "post": {
        "summary": "Skills Gap",
        "operationId": "skills_gap_skills_gap_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SkillsGapRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StandardResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/transition": {
      "post": {
        "summary": "Career Transition",
        "operationId": "career_transition_transition_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TransitionRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StandardResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/education": {
      "post": {
        "summary": "Education",
        "operationId": "education_education_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EducationRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StandardResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/job-creation": {
      "post": {
        "summary": "Job Creation",
        "operationId": "job_creation_job_creation_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/JobCreationRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StandardResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/social-protection": {
      "post": {
        "summary": "Social Protection",
        "operationId": "social_protection_social_protection_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SocialProtectionRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StandardResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/governance": {
      "post": {
        "summary": "Governance",
        "operationId": "governance_governance_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/GovernanceRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StandardResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/equity": {
      "post": {
        "summary": "Equity",
        "operationId": "equity_equity_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EquityRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StandardResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/simulation": {
      "post": {
        "summary": "Simulation",
        "operationId": "simulation_simulation_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SimulationRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StandardResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/dialogue": {
      "post": {
        "summary": "Dialogue",
        "operationId": "dialogue_dialogue_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DialogueRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StandardResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/evaluation": {
      "post": {
        "summary": "Evaluation",
        "operationId": "evaluation_evaluation_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EvaluationRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StandardResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/red-team": {
      "post": {
        "summary": "Red Team",
        "operationId": "red_team_red_team_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RedTeamRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StandardResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/governance/decide": {
      "post": {
        "summary": "Governance Decide",
        "description": "\u0628\u0648\u0627\u0628\u0629 \u0627\u0644\u062d\u0648\u0643\u0645\u0629 \u2014 \u0623\u064a \u0625\u062c\u0631\u0627\u0621 \u062a\u0646\u0641\u064a\u0630\u064a \u064a\u0645\u0631 \u0645\u0646 \u0647\u0646\u0627 \u0642\u0628\u0644 \u062a\u0646\u0641\u064a\u0630\u0647.",
        "operationId": "governance_decide_governance_decide_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/GovernanceApprovalRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DecisionResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/dispatch/{agent_name}": {
      "post": {
        "summary": "Dispatch",
        "description": "\u0645\u0648\u062c\u0651\u0647 \u0639\u0627\u0645 \u2014 \u0644\u0644\u0637\u0628\u0642\u0629 \u0627\u0644\u062a\u0646\u0633\u064a\u0642\u064a\u0629 (orchestration layer).",
        "operationId": "dispatch_dispatch__agent_name__post",
        "parameters": [
          {
            "name": "agent_name",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Agent Name"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "additionalProperties": true,
                "title": "Payload"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StandardResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
...
      "DialogueRequest": {
        "properties": {
          "stakeholder": {
            "type": "string",
            "title": "Stakeholder",
            "description": "employer | worker | union | government | educator | citizen"
          },
          "feedback_text": {
            "type": "string",
            "title": "Feedback Text",
            "description": "\u0646\u0635 \u0627\u0644\u0645\u0644\u0627\u062d\u0638\u0627\u062a \u0645\u0646 \u0627\u0644\u0623\u0637\u0631\u0627\u0641 \u0627\u0644\u0645\u0639\u0646\u064a\u0629"
          }
        },
        "type": "object",
        "required": [
          "stakeholder",
          "feedback_text"
        ],
        "title": "DialogueRequest",
        "description": "\u0648\u0643\u064a\u0644 \u0627\u0644\u062d\u0648\u0627\u0631 \u0627\u0644\u0627\u062c\u062a\u0645\u0627\u0639\u064a \u2014 \u0645\u062e\u0631\u062c\u0647 \u00a78-11."
      },
      "EarlyWarningRequest": {
        "properties": {
          "scope": {
            "type": "string",
            "title": "Scope",
            "description": "national | sector | individual",
            "default": "national"
          },
          "occupation_codes": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Occupation Codes",
            "description": "\u0625\u0646 \u0644\u0645 \u064a\u064f\u0630\u0643\u0631: \u0643\u0644 \u0623\u0637\u0644\u0633 \u0627\u0644\u0645\u0647\u0646"
          },
          "worker": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/WorkerInput"
              },
              {
                "type": "null"
              }
            ],
            "description": "\u0644\u062a\u062d\u0644\u064a\u0644 \u0641\u0631\u062f\u064a"
          }
        },
        "type": "object",
        "title": "EarlyWarningRequest",
        "description": "\u0648\u0643\u064a\u0644 \u0627\u0644\u0625\u0646\u0630\u0627\u0631 \u0627\u0644\u0645\u0628\u0643\u0631 \u2014 \u0645\u062e\u0631\u062c\u0647 \u00a78-1: \u062e\u0631\u064a\u0637\u0629 \u0627\u0644\u0645\u062e\u0627\u0637\u0631 \u0627\u0644\u0648\u0637\u0646\u064a\u0629 + \u0642\u0627\u0626\u0645\u0629 \u0625\u0646\u0630\u0627\u0631\u0627\u062a."
      },
      "EducationRequest": {
        "properties": {
          "institution_type": {
            "type": "string",
            "title": "Institution Type",
            "description": "general | vocational | school",
            "default": "general"
          },
          "target_skills": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Target Skills",
            "description": "\u0627\u0644\u0645\u0647\u0627\u0631\u0627\u062a \u0627\u0644\u0645\u0637\u0644\u0648\u0628\u0629 \u0641\u064a \u0627\u0644\u0633\u0648\u0642"
          },
          "current_curriculum_skills": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Current Curriculum Skills",
            "description": "\u0627\u0644\u0645\u0647\u0627\u0631\u0627\u062a \u0627\u0644\u062a\u064a \u064a\u063a\u0637\u064a\u0647\u0627 \u0627\u0644\u0645\u0646\u0647\u062c"
          },
          "for_occupation": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "For Occupation"
          }
        },
        "type": "object",
        "title": "EducationRequest",
        "description": "\u0648\u0643\u064a\u0644 \u0627\u0644\u062a\u0639\u0644\u064a\u0645 \u2014 \u0645\u062e\u0631\u062c\u0647 \u00a78-5: \u062a\u062d\u0644\u064a\u0644 \u0645\u0646\u0627\u0647\u062c + \u0648\u062d\u062f\u0627\u062a \u062a\u062f\u0631\u064a\u0628 \u0642\u0635\u064a\u0631\u0629."
      },
      "EquityRequest": {
        "properties": {
          "population": {
            "items": {
              "$ref": "#/components/schemas/WorkerInput"
            },
            "type": "array",
            "title": "Population",
            "description": "\u0639\u064a\u0646\u0629 \u0633\u0643\u0627\u0646\u064a\u0629 \u0644\u0644\u062a\u062d\u0644\u064a\u0644"
          }
        },
        "type": "object",
        "required": [
          "population"
        ],
        "title": "EquityRequest",
        "description": "\u0648\u0643\u064a\u0644 \u0627\u0644\u0639\u062f\u0627\u0644\u0629 \u2014 \u0645\u062e\u0631\u062c\u0647 \u00a78-9."
      },
      "EvaluationRequest": {
        "properties": {
          "kpi_code": {
            "type": "string",
            "title": "Kpi Code",
            "description": "transition_completion_rate | six_month_placement | wage_change | gender_gap_closure | rural_access"
          },
          "measured_value": {
            "type": "number",
            "title": "Measured Value"
          },
          "target_value": {
            "type": "number",
            "title": "Target Value"
          },
          "period": {
            "type": "string",
            "title": "Period",
            "description": "current | previous",
            "default": "current"
          }
        },
        "type": "object",
        "required": [
          "kpi_code",
          "measured_value",
          "target_value"
        ],
        "title": "EvaluationRequest",
        "description": "\u0648\u0643\u064a\u0644 \u0627\u0644\u062a\u0642\u064a\u064a\u0645 \u0648\u0627\u0644\u0645\u0633\u0627\u0621\u0644\u0629 \u2014 \u0645\u062e\u0631\u062c\u0647 \u00a78-12."
      },
      "ExposureRequest": {
        "properties": {
          "occupation_code": {
            "type": "string",
            "title": "Occupation Code"
          },
          "task_breakdown": {
            "anyOf": [
              {
                "items": {
                  "additionalProperties": {
                    "type": "string"
                  },
                  "type": "object"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Task Breakdown",
            "description": "\u0645\u0647\u0627\u0645 \u0645\u062e\u0635\u0635\u0629 {label_ar, label_en, category, time_fraction} \u0644\u062a\u0642\u064a\u064a\u0645 \u0645\u0647\u0646\u0629 \u063a\u064a\u0631 \u0645\u0633\u062c\u0644\u0629"
          }
        },
        "type": "object",
        "required": [
          "occupation_code"
        ],
        "title": "ExposureRequest",
        "description": "\u0648\u0643\u064a\u0644 \u0627\u0644\u062a\u0639\u0631\u0636 \u0627\u0644\u0645\u0647\u0646\u064a \u2014 \u0645\u062e\u0631\u062c\u0647 \u00a78-2."
      },
...
      "GovernanceRequest": {
        "properties": {
          "system_type": {
            "type": "string",
            "title": "System Type",
            "description": "\u0646\u0648\u0639 \u0627\u0644\u0646\u0638\u0627\u0645 \u0627\u0644\u062e\u0648\u0627\u0631\u0632\u0645\u064a: hiring | worker_management | promotion | other"
          },
          "decision_description": {
            "type": "string",
            "title": "Decision Description",
            "description": "\u0648\u0635\u0641 \u0627\u0644\u0642\u0631\u0627\u0631 \u0627\u0644\u0630\u064a \u064a\u062a\u062e\u0630\u0647 \u0627\u0644\u0646\u0638\u0627\u0645"
          },
          "affected_groups": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Affected Groups",
            "description": "\u0627\u0644\u0641\u0626\u0627\u062a \u0627\u0644\u0645\u062a\u0623\u062b\u0631\u0629: young | women | disabled | rural | low_income"
          }
        },
        "type": "object",
        "required": [
          "system_type",
          "decision_description"
        ],
        "title": "GovernanceRequest",
        "description": "\u0648\u0643\u064a\u0644 \u0627\u0644\u062d\u0648\u0643\u0645\u0629 \u0648\u0627\u0644\u0627\u0645\u062a\u062b\u0627\u0644 \u2014 \u0645\u062e\u0631\u062c\u0647 \u00a78-8."
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "JobCreationRequest": {
        "properties": {
          "region": {
            "type": "string",
            "title": "Region",
            "default": ""
          },
          "sectors": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Sectors",
            "description": "\u0623\u0643\u0648\u0627\u062f \u0627\u0644\u0642\u0637\u0627\u0639\u0627\u062a \u0645\u0646 GROWTH_SECTORS \u0623\u0648 \u0643\u0644 \u0627\u0644\u0642\u0637\u0627\u0639\u0627\u062a"
          }
        },
        "type": "object",
        "title": "JobCreationRequest",
        "description": "\u0648\u0643\u064a\u0644 \u062e\u0644\u0642 \u0627\u0644\u0648\u0638\u0627\u0626\u0641 \u2014 \u0645\u062e\u0631\u062c\u0647 \u00a78-6."
      },
...
      "SimulationRequest": {
        "properties": {
          "policy_name": {
            "type": "string",
            "title": "Policy Name",
            "description": "\u0627\u0633\u0645 \u0627\u0644\u0633\u064a\u0627\u0633\u0629 \u0645\u062d\u0644 \u0627\u0644\u0645\u062d\u0627\u0643\u0627\u0629"
          },
          "displacement_rate": {
            "type": "number",
            "maximum": 1.0,
            "minimum": 0.0,
            "title": "Displacement Rate",
            "description": "\u0645\u0639\u062f\u0644 \u0625\u0632\u0627\u062d\u0629 \u0627\u0644\u0639\u0645\u0627\u0644\u0629 \u0641\u064a \u0627\u0644\u0645\u0647\u0627\u0645 \u0627\u0644\u0622\u0644\u064a\u0629"
          },
          "reinstatement_rate": {
            "type": "number",
            "maximum": 1.0,
            "minimum": 0.0,
            "title": "Reinstatement Rate",
            "description": "\u0645\u0639\u062f\u0644 \u0625\u0639\u0627\u062f\u0629 \u0627\u0644\u062a\u0645\u0643\u064a\u0646 \u0639\u0628\u0631 \u0627\u0644\u0645\u0647\u0627\u0645 \u0627\u0644\u062c\u062f\u064a\u062f\u0629"
          },
          "horizon_years": {
            "type": "integer",
            "maximum": 20.0,
            "minimum": 1.0,
            "title": "Horizon Years",
            "default": 5
          },
          "sector_codes": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Sector Codes"
          }
        },
        "type": "object",
        "required": [
          "policy_name",
          "displacement_rate",
          "reinstatement_rate"
        ],
        "title": "SimulationRequest",
        "description": "\u0648\u0643\u064a\u0644 \u0627\u0644\u0645\u062d\u0627\u0643\u0627\u0629 \u2014 \u0645\u062e\u0631\u062c\u0647 \u00a78-10: \u0633\u064a\u0646\u0627\u0631\u064a\u0648\u0647\u0627\u062a Acemoglu\u2013Restrepo (\u0627\u0633\u062a\u0628\u062f\u0627\u0644/\u0627\u0633\u062a\u062d\u062f\u0627\u062b)."
      },
      "SkillsGapRequest": {
        "properties": {
          "worker": {
            "$ref": "#/components/schemas/WorkerInput"
          },
          "target_occupation": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Target Occupation",
            "description": "\u0625\u0646 \u0644\u0645 \u064a\u064f\u062d\u062f\u062f: \u062a\u064f\u0633\u062a\u0646\u062a\u062c \u0645\u0646 \u0623\u0641\u0636\u0644 \u0645\u0633\u0627\u0631 \u0627\u0646\u062a\u0642\u0627\u0644"
          }
        },
        "type": "object",
        "required": [
          "worker"
        ],
        "title": "SkillsGapRequest",
        "description": "\u0648\u0643\u064a\u0644 \u0641\u062c\u0648\u0629 \u0627\u0644\u0645\u0647\u0627\u0631\u0627\u062a \u2014 \u0645\u062e\u0631\u062c\u0647 \u00a78-3."
      },
      "SocialProtectionRequest": {
        "properties": {
          "worker": {
            "$ref": "#/components/schemas/WorkerInput"
          },
          "transition_horizon_months": {
            "type": "integer",
            "maximum": 36.0,
            "minimum": 1.0,
            "title": "Transition Horizon Months",
            "default": 12
          }
        },
        "type": "object",
        "required": [
          "worker"
        ],
        "title": "SocialProtectionRequest",
        "description": "\u0648\u0643\u064a\u0644 \u0627\u0644\u062d\u0645\u0627\u064a\u0629 \u0627\u0644\u0627\u062c\u062a\u0645\u0627\u0639\u064a\u0629 \u2014 \u0645\u062e\u0631\u062c\u0647 \u00a78-7."
      },
      "StandardResponse": {
        "properties": {
          "status": {
            "type": "string",
            "title": "Status",
            "default": "ok"
          },
          "agent": {
            "type": "string",
            "title": "Agent"
          },
          "result": {
            "additionalProperties": true,
            "type": "object",
            "title": "Result"
          },
          "governance": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Governance",
            "description": "\u0642\u0631\u0627\u0631 \u0628\u0648\u0627\u0628\u0629 \u0627\u0644\u062d\u0648\u0643\u0645\u0629 \u0625\u0646 \u0648\u064f\u062c\u062f"
          }
        },
        "type": "object",
        "required": [
          "agent",
          "result"
        ],
        "title": "StandardResponse"
      },
      "TransitionRequest": {
        "properties": {
          "worker": {
            "$ref": "#/components/schemas/WorkerInput"
          },
          "target_occupation": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Target Occupation"
          },
          "max_paths": {
            "type": "integer",
            "maximum": 10.0,
            "minimum": 1.0,
            "title": "Max Paths",
            "default": 3
          }
        },
        "type": "object",
        "required": [
          "worker"
        ],
        "title": "TransitionRequest",
        "description": "\u0648\u0643\u064a\u0644 \u0627\u0644\u0627\u0646\u062a\u0642\u0627\u0644 \u0627\u0644\u0645\u0647\u0646\u064a \u2014 \u0645\u062e\u0631\u062c\u0647 \u00a78-4: \u062e\u0637\u0629 \u0627\u0646\u062a\u0642\u0627\u0644 \u0641\u0631\u062f\u064a\u0629 \u0643\u0627\u0645\u0644\u0629."
      },
...
    }
  }
}